Refactor parser to allow modular parsing methods
This patch is huge. It basically allows to write individual parsing
functions (like parse_attributes, parse_function) to help modularize
parsing. Lexer member in Parser type is now boxed, but may be lifted one
more time to check compatibility with Rc<RefCell<Lexer>>.

Also error reporting is messed up because of it. That needs to be done.

It also has some utility functions to help in lexing.