Fix correct error reporting for parse_import
Expected function name and semicolon were reported with wrong location,
because they were being passed as only error kinds and without location
info. But as in this particular case, error reporting is handled inside
the parse_import and not up the call stack, so needed to specifically
return the correct location info.