Add qubit v/s bit rule in type checker
This rule checks a let assignment for a qubit when assigned a bit and
vice versa. Both scenarios are considered fine accorindg to our
subtyping rules, so the type checker ignores any mismatch. This is
because the codegen will auto-magically put required stubs prior to
these instructions. More details are put in the comment section with
examples.