Remove explicit types from examples
The type inference will handle this. Although there are bugs in the type
checker, like putting a type signature for `choice` will actually result
in a type mismatch (it shouldn't).

Also, definition of Hadamard is not written. It would require a few
things working before we jump on it. So right now it is an identity
operator.