Enhance toss example and add type constriant between qubit and float
Now the toss example in `examples/` is feature complete. For small
values of angles (for sin and cos approx.) it is a standalone code to
demonstrate the tossing of a coin.

The type constaint for which a qubit when operated on by a floating
value returns a qubit, is added. This is needed for toss example when
the unitary transformation is applied.