Add type inference for conditionals
This supports the conditionals themselves being treated as expressions.
So a conditional block now can be used in let expressions.

    let x = if conditional { return foo(); } else { return bar(); }

It returns the type of its last returning expression (ending type). The
type checker also checks if all ending types in each if/else if/else
blocks is same. That is, in above snippet, foo := bar.