Add support for 'alias' keyword
Synatax:
  alias Foo = function;

Any function, outside of the current module, when imported via import,
can be aliased to other names via 'alias' keyword. The parser resolves
it by adding a stub function with the name 'Foo' and doing a function
call to 'function'. The stub is added in the current module, where alias
is assigned.

This is first phase of adding alias support. This does not parse
correctly and adds a stub for aliases for functions defined in the same
module.